requirements.txt
fastapi==0.111.0
starlette==0.37.2
pydantic>=2.11.0
mangum==0.17.0
uvicorn==0.30.0
python-multipart==0.0.9
aws-lambda-powertools==3.0.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-genai>=1.0.0